+++ src/main/java/kitchenpos/application/TableService.java
@@ -4,10 +4,10 @@ import kitchenpos.domain.order.OrderRepository;
 import kitchenpos.domain.order.OrderStatus;
 import kitchenpos.domain.order.OrderTable;
 import kitchenpos.domain.order.OrderTableRepository;
-import kitchenpos.ui.dto.OrderTableResponse;
-import kitchenpos.ui.dto.OrderTableEmptyRequest;
-import kitchenpos.ui.dto.OrderTableGuestRequest;
-import kitchenpos.ui.dto.OrderTableRequest;
+import kitchenpos.ui.dto.order.OrderTableResponse;
+import kitchenpos.ui.dto.order.OrderTableEmptyRequest;
+import kitchenpos.ui.dto.order.OrderTableGuestRequest;
+import kitchenpos.ui.dto.order.OrderTableRequest;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
