+++ src/main/java/kitchenpos/ui/dto/order/OrderResponse.java
@@ -1,4 +1,4 @@
-package kitchenpos.ui.dto;
+package kitchenpos.ui.dto.order;
 
 import kitchenpos.domain.order.Order;
