+++ src/main/java/kitchenpos/ui/dto/order/OrderTableEmptyRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.ui.dto;
+package kitchenpos.ui.dto.order;
 
 public class OrderTableEmptyRequest {
