+++ src/main/java/kitchenpos/ui/dto/order/OrderTableGuestRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.ui.dto;
+package kitchenpos.ui.dto.order;
 
 import javax.validation.constraints.Positive;
