+++ src/test/java/kitchenpos/application/ProductServiceTest.java
@@ -1,8 +1,8 @@
 package kitchenpos.application;
 
 import kitchenpos.SpringBootTestSupport;
-import kitchenpos.ui.dto.ProductRequest;
-import kitchenpos.ui.dto.ProductResponse;
+import kitchenpos.ui.dto.product.ProductRequest;
+import kitchenpos.ui.dto.product.ProductResponse;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
