+++ src/test/java/kitchenpos/application/TableServiceTest.java
@@ -7,10 +7,10 @@ import kitchenpos.domain.order.OrderStatus;
 import kitchenpos.domain.order.OrderTable;
 import kitchenpos.domain.order.TableGroup;
 import kitchenpos.domain.product.Product;
-import kitchenpos.ui.dto.OrderTableResponse;
-import kitchenpos.ui.dto.OrderTableEmptyRequest;
-import kitchenpos.ui.dto.OrderTableGuestRequest;
-import kitchenpos.ui.dto.OrderTableRequest;
+import kitchenpos.ui.dto.order.OrderTableResponse;
+import kitchenpos.ui.dto.order.OrderTableEmptyRequest;
+import kitchenpos.ui.dto.order.OrderTableGuestRequest;
+import kitchenpos.ui.dto.order.OrderTableRequest;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
