+++ src/test/java/kitchenpos/ui/OrderRestControllerTest.java
@@ -2,7 +2,7 @@ package kitchenpos.ui;
 
 import kitchenpos.application.OrderService;
 import kitchenpos.domain.order.OrderStatus;
-import kitchenpos.ui.dto.*;
+import kitchenpos.ui.dto.order.*;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
